casting size_t arguments to int to deal with the storage size difference on 64-bit architectures

+++ src/examples/file_transfer/client.c
@@ -8,6 +8,9 @@
 /****************************************************************************/
 /* $Id$
  * $Log$
+ * Revision 1.11  2005/11/09 18:39:33  alsu
+ * casting size_t arguments to int to deal with the storage size difference on 64-bit architectures
+ *
  * Revision 1.10  2005/10/05 14:02:29  alsu
  * eliminating gcc warning about signedness
  *
@@ -55,7 +58,7 @@ main(int argc, char* argv[])
   diet_profile_t* profile = NULL;
   int *size1 = NULL;
   int *size2 = NULL;
-  unsigned int out_size = 0;
+  size_t out_size = 0;
 
 
   if (argc != 4) {
@@ -98,7 +101,7 @@ main(int argc, char* argv[])
     diet_file_get(diet_parameter(profile,4), NULL, &out_size, &path);
     if (path && (*path != '\0')) {
       printf("Location of returned file is %s, its size is %d.\n",
-	     path, out_size);
+	     path, (int) out_size);
       diet_free_data(diet_parameter(profile,4));
     }
   }
